requirements.txt
scipy==1.1.0
tensorflow_gpu==1.14.0
six==1.11.0
sympy==1.3
numpy==1.15.4
scikit_learn==0.21.3
jupyter==1.0.0
